requirements.txt
PyQt6==6.6.1
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
reportlab==4.0.7
pyqtgraph==0.13.7
python-dotenv==1.0.0
uuid==1.30
